backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
redis==5.2.0
arq==0.26.1
httpx==0.28.0
python-multipart==0.0.12
